Refactor: Rename Task component
